+++ src/index.css
@@ -1,0 +1,23 @@
+html {
+  box-sizing: border-box;
+}
+
+*,
+*:after,
+*:before {
+  box-sizing: inherit;
+}
+
+h1 {
+  font-family: 'Signika Negative', sans-serif;
+}
+
+blockquote {
+  border-left: 10px solid rgba(204, 204, 204, 0.67);
+}
+
+blockquote::before {
+  content: open-quote;
+  color: rgb(155, 155, 155);
+  font-size: 1.5em;
+}
